backend/course/urls.py
from django.urls import path, include
from rest_framework_nested import routers
from . import views


app_name = 'course'

urlpatterns = [

    # Tag URLs
    path('tags/', views.TagListCreateView.as_view(), name='tag-list'),
    path('tags/<slug:slug>/', views.TagDetailView.as_view(), name='tag-detail'),
    path('my-courses/', views.UserCoursesView.as_view(), name='user-courses'),

    # Course URLs
    path('courses/', views.CourseListCreateView.as_view(), name='course-list'),
    path('courses/<slug:slug>/enroll/', views.CourseEnrollView.as_view(), name='course-enroll'),
    path('courses/<slug:slug>/rate/', views.CourseRateView.as_view(), name='course-rate'),

    path('courses/<slug:slug>/', views.CourseDetailView.as_view(), name='course-detail'),

    # Chapter URLs
    # path('courses/<slug:course_slug>/chapters/', views.ChapterListCreateView.as_view(), name='chapter-list'),
    # path('courses/<slug:course_slug>/chapters/<int:pk>/', views.ChapterDetailView.as_view(), name='chapter-detail'),

    # Lesson URLs
    # path('courses/<slug:course_slug>/chapters/<int:chapter_pk>/lessons/', views.LessonListCreateView.as_view(), name='lesson-list'),
    # path('courses/<slug:course_slug>/chapters/<int:chapter_pk>/lessons/<int:pk>/', views.LessonDetailView.as_view(), name='lesson-detail'),
    # path('courses/<slug:course_slug>/chapters/<int:chapter_pk>/lessons/<int:pk>/mark-complete/', views.LessonMarkCompleteView.as_view(), name='lesson-mark-complete'),
    # path('courses/<slug:course_slug>/chapters/<int:chapter_pk>/lessons/<int:pk>/update-progress/', views.LessonUpdateProgressView.as_view(), name='lesson-update-progress'),

    # Enrollment URLs
    # path('enrollments/', views.EnrollmentListV/iew.as_view(), name='enrollment-list'),

    # Discount URLs
    # path('discounts/', views.DiscountListCreateView.as_view(), name='discount-list'),
    # path('discounts/<int:pk>/', views.DiscountDetailView.as_view(), name='discount-detail'),
    # path('discounts/<int:pk>/validate/', views.DiscountValidateView.as_view(), name='discount-validate'),
    #
    # # LessonProgress URLs
    # path('lesson-progress/', views.LessonProgressListCreateView.as_view(), name='lesson-progress-list'),
    # path('lesson-progress/<int:pk>/', views.LessonProgressDetailView.as_view(), name='lesson-progress-detail'),
    # path('lesson-progress/<int:pk>/mark-completed/', views.LessonProgressMarkCompletedView.as_view(), name='lesson-progress-mark-completed'),
    # path('lesson-progress/<int:pk>/update-progress/', views.LessonProgressUpdateProgressView.as_view(), name='lesson-progress-update-progress'),

    # CourseProgress URL
    path('courses/<int:course_pk>/progress/', views.CourseProgressView.as_view(), name='course-progress'),

    # ChapterProgress URL
    path('courses/<int:course_pk>/chapter-progress/', views.ChapterProgressView.as_view(), name='chapter-progress'),

    # EnrollmentProgress URL
    path('courses/<int:course_pk>/enrollment-progress/', views.EnrollmentProgressView.as_view(), name='enrollment-progress'),

    # Learn Page URLs
    path('courses/<int:pk>/learn/', views.CourseLearnView.as_view(), name='course-learn'),
    path('lessons/<int:pk>/progress/', views.LessonProgressUpdateView.as_view(), name='lesson-progress-update'),
    path('lessons/<int:pk>/notes/', views.LessonNoteView.as_view(), name='lesson-notes'),
    path('lessons/<int:pk>/bookmarks/', views.LessonBookmarkView.as_view(), name='lesson-bookmarks'),

    # User Courses URL
]